Refactor of the project after last merge
