+++ app/src/androidTest/java/ch/epfl/favo/MainActivityTest.java
@@ -7,17 +7,18 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import ch.epfl.favo.testhelpers.FakeFirebaseUser;
 import ch.epfl.favo.util.DependencyFactory;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.swipeLeft;
-
+import static androidx.test.espresso.action.ViewActions.swipeRight;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
-import static androidx.test.espresso.matcher.ViewMatchers.withText;
-import static ch.epfl.favo.testhelpers.TestConstants.*;
+import static ch.epfl.favo.TestConstants.EMAIL;
+import static ch.epfl.favo.TestConstants.NAME;
+import static ch.epfl.favo.TestConstants.PHOTO_URI;
+import static ch.epfl.favo.TestConstants.PROVIDER;
 
 @RunWith(AndroidJUnit4.class)
 public class MainActivityTest {
